refactor: rename getAuthUserDetails to getAuthDetails
