+++ src/utils/index.js
@@ -1,3 +1,4 @@
-export { getUserDetails } from "./getAuthUserDetails";
+export { getUserDetails } from "./getUserDetails";
 export { getUnfollowedUsers } from "./getUnfollowedUsers";
 export { getUnfollowedUsersPosts } from "./getUnfollowedUsers";
+export { getSortedPosts } from "./getSortedPosts";
